+++ Pulsar/src/RendererSettings.h
@@ -9,7 +9,7 @@ namespace _RendererSettings
 {
 	extern const char* settings_filepath;
 	extern std::string rect_renderable_filepath;
-	extern float initial_window_width, initial_window_height;
+	extern int initial_window_width, initial_window_height;
 	extern TextureSlot max_texture_slots;
 	extern VertexSize standard_vertex_pool_size;
 	extern VertexSize standard_index_pool_size;
